explode useMapLayers logic into different files, one for each layer family
