+++ components/map/useMapLayers.ts
@@ -1,138 +1,33 @@
-// Background styles
-import vectorOsm from '@/components/map/mapstyles/vector-osm.json';
-import vectorIgnStandard from '@/components/map/mapstyles/vector-ign-standard.json';
-import satellite from '@/components/map/mapstyles/satellite.json';
-import satellite_2016_2020 from '@/components/map/mapstyles/satellite_2016_2020.json';
-
 // Maplibre styles
-import maplibregl, {
-  FilterSpecification,
-  StyleSpecification,
-} from 'maplibre-gl';
+import maplibregl from 'maplibre-gl';
 
 // React things
-import { useCallback, useEffect, useRef } from 'react';
+import { useEffect, useRef } from 'react';
 
 // Store
 import { useDispatch, useSelector } from 'react-redux';
 import { Actions, AppDispatch, RootState } from '@/stores/store';
-
-// Images
-import reportIcon from '@/public/images/map/report.png';
-import checkGreenIcon from '@/public/images/map/check-green.png';
-
-///////////////////////////////////
-///////////////////////////////////
-// BUILDINGS
-
-// Buildings source
-export const SRC_BDGS_POINTS = 'bdgtiles_points';
-export const SRC_BDGS_SHAPES = 'bdgtiles_shapes';
-export const SRC_BDGS_SHAPES_URL = `${process.env.NEXT_PUBLIC_API_BASE}/tiles/shapes/{x}/{y}/{z}.pbf?only_active_and_real=false`;
-export const SRC_BDGS_POINTS_URL = `${process.env.NEXT_PUBLIC_API_BASE}/tiles/{x}/{y}/{z}.pbf?only_active_and_real=false`;
-
-// Buildings layers : point:
-export const LAYER_BDGS_POINT = 'bdgs_point';
-export const LAYER_BDGS_POINT_SHAPE_BORDER = 'bdgs_bdgs_point_shape_border';
-export const LAYER_BDGS_POINT_SHAPE_FILL = 'bdgs_bdgs_point_shape_fill';
-export const LAYER_BDGS_POINT_SHAPE_VERIFIED_CHECK =
-  'bdgs_point_shape_marked_hatch';
-export const LAYERS_BDGS_POINT_ALL = [
-  LAYER_BDGS_POINT,
-  LAYER_BDGS_POINT_SHAPE_BORDER,
-  LAYER_BDGS_POINT_SHAPE_FILL,
-  LAYER_BDGS_POINT_SHAPE_VERIFIED_CHECK,
-];
-
-// Buildings layers : shapes
-export const LAYER_BDGS_SHAPE_BORDER = 'bdgs_shape';
-export const LAYER_BDGS_SHAPE_FILL = 'bdgs_shape_fill';
-export const LAYER_BDGS_SHAPE_POINT = 'bdgs_shape_point';
-export const LAYER_BDGS_SHAPE_MARKED_HATCH = 'bdgs_shape_marked_hatch';
-export const LAYERS_BDGS_SHAPE_ALL = [
-  LAYER_BDGS_SHAPE_BORDER,
-  LAYER_BDGS_SHAPE_FILL,
-  LAYER_BDGS_SHAPE_POINT,
-  LAYER_BDGS_SHAPE_MARKED_HATCH,
-];
-
-////////////////////////////////////
-////////////////////////////////////
-// Reports
-export const SRC_REPORTS = 'reports';
-export const LAYER_REPORTS_CIRCLE = 'reports_circle';
-export const LAYER_REPORTS_ICON = 'reports_icon';
-export const LAYER_REPORTS_SMALL_CIRCLES = 'report_small_circles';
-
-export const SRC_REPORTS_URL = `${process.env.NEXT_PUBLIC_API_BASE}/reports/tiles/{x}/{y}/{z}.pbf`;
-
-////////////////////////////////////
-////////////////////////////////////
-// BAN Adresses
-
-// BAN source
-export const SRC_BAN = 'ban';
-export const SRC_BAN_URL = `https://plateforme.adresse.data.gouv.fr/tiles/ban/{z}/{x}/{y}.pbf`;
-
-// BAN layer
-export const LAYER_BAN_POINT = 'ban_points';
-export const LAYER_BAN_TXT = 'ban_txt';
-
-///////////////////////////////////
-///////////////////////////////////
-// ADS
-
-// ADS source
-export const SRC_ADS = 'ads';
-export const SRC_ADS_URL = `${process.env.NEXT_PUBLIC_API_BASE}/permis/tiles/{x}/{y}/{z}.pbf`;
-
-// ADS Layers
-export const LAYER_ADS_CIRCLE = 'adscircle';
-export const LAYER_ADS_ICON = 'adsicon';
-
-////////////////////////////////////
-////////////////////////////////////
-// Plots
-export const LAYER_PLOTS_SHAPE = 'plots_shape';
-export const LAYER_PLOTS_TXT = 'plots_txt';
-export const SRC_PLOTS = 'plotstiles';
-
-// Icons
-import { getADSOperationIcons } from '@/logic/ads';
 import {
   MapBackgroundLayer,
   MapBuildingsLayer,
   MapExtraLayer,
 } from '@/stores/map/map-slice';
-import { setDisplayedReportFilters } from './report/useMapStateSyncReport';
 
-export const STYLES: Record<
-  MapBackgroundLayer,
-  { name: string; style: StyleSpecification }
-> = {
-  vectorOsm: {
-    name: 'OSM',
-    style: vectorOsm as StyleSpecification,
-  },
-  vectorIgnStandard: {
-    name: 'IGN standard',
-    style: vectorIgnStandard as StyleSpecification,
-  },
-  satellite: {
-    name: 'Satellite',
-    style: satellite as StyleSpecification,
-  },
-  satellite_2016_2020: {
-    name: 'Satellite 2011-2016',
-    style: satellite_2016_2020 as StyleSpecification,
-  },
-};
+// Map layers (one module per map element)
+import { STYLES } from './layers/backgrounds';
+import { installBuildings } from './layers/buildings';
+import { installADS } from './layers/ads';
+import { installBAN } from './layers/ban';
+import { installPlots } from './layers/plots';
+import { installReports } from './layers/reports';
 
-export const DEFAULT_STYLE =
-  STYLES[
-    (process.env.FOND_DE_CARTE_PAR_DEFAUT as keyof typeof STYLES) ||
-      'vectorIgnStandard'
-  ].style;
+// Re-export the constants and helpers so existing imports from this module keep working
+export { STYLES, DEFAULT_STYLE } from './layers/backgrounds';
+export * from './layers/buildings';
+export * from './layers/ads';
+export * from './layers/ban';
+export * from './layers/plots';
+export * from './layers/reports';
 
 // C.f. discussion here https://github.com/mapbox/mapbox-gl-js/issues/6707#issuecomment-1942879968
 // and https://github.com/mapbox/mapbox-gl-draw/blob/main/src/setup.js#L66
@@ -201,7 +96,7 @@ export const useMapLayers = ({
       }
 
       if (layers.extraLayers.includes('reports')) {
-        installReports(map);
+        await installReports(map, displayedReportTags);
       }
     } catch (e) {
       throw e;
@@ -216,611 +111,6 @@ export const useMapLayers = ({
       });
     }
   };
-
-  ///////////////////////////////////
-  ///////////////////////////////////
-  // ADS
-
-  const installADS = useCallback(async (map: maplibregl.Map) => {
-    if (map.getLayer(LAYER_ADS_CIRCLE)) map.removeLayer(LAYER_ADS_CIRCLE);
-    if (map.getLayer(LAYER_ADS_ICON)) map.removeLayer(LAYER_ADS_ICON);
-    if (map.getSource(SRC_ADS)) map.removeSource(SRC_ADS);
-
-    // Icons for ADS
-    // build icon
-    if (!map.hasImage('adsBuild')) {
-      const adsBuild = await map.loadImage(adsOperationsIcons.build.src);
-      map.addImage('adsBuild', adsBuild.data, { sdf: true });
-    }
-
-    // modify icon
-    if (!map.hasImage('adsModify')) {
-      const adsModify = await map.loadImage(adsOperationsIcons.modify.src);
-      map.addImage('adsModify', adsModify.data, { sdf: true });
-    }
-
-    // demolish icon
-    if (!map.hasImage('adsDemo')) {
-      const adsDemo = await map.loadImage(adsOperationsIcons.demolish.src);
-      map.addImage('adsDemo', adsDemo.data, { sdf: true });
-    }
-
-    map.addSource(SRC_ADS, {
-      type: 'vector',
-      tiles: [SRC_ADS_URL + '#' + Math.random()],
-      minzoom: 16,
-      maxzoom: 22,
-      promoteId: 'file_number',
-    });
-
-    map.addLayer({
-      id: LAYER_ADS_CIRCLE,
-      source: SRC_ADS,
-      'source-layer': 'default',
-      type: 'circle',
-      paint: {
-        'circle-radius': [
-          'case',
-          ['boolean', ['==', ['feature-state', 'hovered'], true]],
-          12,
-          10,
-        ],
-        'circle-stroke-color': [
-          'case',
-          ['boolean', ['feature-state', 'in_panel'], false],
-          '#431407',
-          '#ffffff',
-        ],
-        'circle-stroke-width': 2,
-        'circle-color': '#fbbf24',
-      },
-    });
-
-    map.addLayer({
-      id: LAYER_ADS_ICON,
-      source: SRC_ADS,
-      'source-layer': 'default',
-      type: 'symbol',
-      layout: {
-        'icon-image': [
-          'match',
-          ['get', 'operation'],
-          'build',
-          'adsBuild',
-          'demolish',
-          'adsDemo',
-          'modify',
-          'adsModify',
-          'adsBuild',
-        ],
-        'icon-size': 0.2,
-        'icon-allow-overlap': true,
-        'icon-ignore-placement': true,
-      },
-      paint: {
-        'icon-color': '#431407',
-      },
-    });
-  }, []);
-
-  ///////////////////////////////////
-  ///////////////////////////////////
-  // Buildings
-  const installBuildings = (map: maplibregl.Map) => {
-    // First, we remove buildings layers and source
-    removeBuildings(map);
-
-    // Then, we install the source
-    installBuildingsSource(map);
-    installBuildingsLayers(map);
-  };
-
-  const installBuildingsSource = (map: maplibregl.Map) => {
-    map.addSource(SRC_BDGS_POINTS, {
-      type: 'vector',
-      tiles: [SRC_BDGS_POINTS_URL],
-      minzoom: 16,
-      maxzoom: 22,
-      promoteId: 'rnb_id',
-    });
-
-    map.addSource(SRC_BDGS_SHAPES, {
-      type: 'vector',
-      // the random element is used to bypass the browser caching
-      // when a building's shape is updated, we change the random element
-      // to force a clean re-render of the data on the map
-      tiles: [SRC_BDGS_SHAPES_URL + '#' + Math.random()],
-      minzoom: 16,
-      maxzoom: 22,
-      promoteId: 'rnb_id',
-    });
-  };
-
-  const getDefaultBuildingFeatureFilter = () => {
-    const defaultBuildingFeatureFilter: any = [
-      'all',
-      ['==', 'is_active', true],
-      ['!=', 'status', 'demolished'],
-    ];
-
-    return defaultBuildingFeatureFilter;
-  };
-
-  const installBuildingsLayers = (map: maplibregl.Map) => {
-    if (layers.buildings == 'point') {
-      installBuildingsPointsLayers(map);
-    }
-    if (layers.buildings == 'polygon') {
-      installBuildingsShapesLayers(map);
-    }
-  };
-
-  const installBuildingsPointsLayers = async (map: maplibregl.Map) => {
-    const defaultBuildingFeatureFilter = getDefaultBuildingFeatureFilter();
-    const verifiedActive = layers.extraLayers.includes('verified');
-
-    if (!map.hasImage('checkGreen')) {
-      const check = await map.loadImage(checkGreenIcon.src);
-      if (!map.hasImage('checkGreen')) {
-        map.addImage('checkGreen', check.data);
-      }
-    }
-
-    // Building shapes for vector background
-    if (['vectorIgnStandard', 'vectorOsm'].includes(layers.background)) {
-      map.addLayer({
-        id: LAYER_BDGS_POINT_SHAPE_FILL,
-        type: 'fill',
-        filter: defaultBuildingFeatureFilter,
-        source: SRC_BDGS_SHAPES,
-        'source-layer': 'default',
-        paint: {
-          'fill-color': '#dcdcdc',
-          'fill-opacity': 1,
-        },
-      });
-
-      // add check marks in the consultation map
-      map.addLayer({
-        id: LAYER_BDGS_POINT_SHAPE_VERIFIED_CHECK,
-        type: 'fill',
-        source: SRC_BDGS_SHAPES,
-        'source-layer': 'default',
-        filter: [
-          'all',
-          ...defaultBuildingFeatureFilter.slice(1),
-          ['==', 'is_marked_as_correct', true],
-        ],
-        paint: {
-          'fill-pattern': 'checkGreen',
-          'fill-opacity': 0.9,
-        },
-      });
-
-      map.addLayer({
-        id: LAYER_BDGS_POINT_SHAPE_BORDER,
-        type: 'line',
-        filter: defaultBuildingFeatureFilter,
-        source: SRC_BDGS_SHAPES,
-        'source-layer': 'default',
-        paint: {
-          'line-color': [
-            'case',
-            ['boolean', ['feature-state', 'in_panel'], false],
-            '#31e060',
-            '#00000033',
-          ],
-          'line-width': [
-            'case',
-            ['boolean', ['feature-state', 'in_panel'], false],
-            3,
-            1.5,
-          ],
-        },
-      });
-    }
-
-    map.addLayer({
-      id: LAYER_BDGS_POINT,
-      type: 'circle',
-      source: SRC_BDGS_POINTS,
-      'source-layer': 'default',
-      filter: defaultBuildingFeatureFilter,
-      paint: {
-        'circle-radius': [
-          'case',
-          ['boolean', ['==', ['feature-state', 'hovered'], true]],
-          6,
-          5,
-        ],
-        'circle-stroke-color': '#ffffff',
-        'circle-stroke-width': 3,
-        'circle-color': [
-          'case',
-          ['boolean', ['feature-state', 'in_panel'], false],
-          '#31e060',
-          '#1452e3',
-        ],
-      },
-    });
-  };
-
-  const installBuildingsShapesLayers = async (map: maplibregl.Map) => {
-    const defaultBuildingFeatureFilter = getDefaultBuildingFeatureFilter();
-    const selectedBuildingColor = selectedBuildingisGreen
-      ? '#31e060'
-      : '#1452e3';
-    const verifiedActive = layers.extraLayers.includes('verified');
-    const isSatellite = ['satellite', 'satellite_2016_2020'].includes(
-      layers.background,
-    );
-
-    if (verifiedActive && !map.hasImage('hatchGreen')) {
-      const hatch = await map.loadImage(checkGreenIcon.src);
-      if (!map.hasImage('hatchGreen')) {
-        map.addImage('hatchGreen', hatch.data);
-      }
-    }
-
-    map.addLayer({
-      id: LAYER_BDGS_SHAPE_BORDER,
-      type: 'fill',
-      source: SRC_BDGS_SHAPES,
-      'source-layer': 'default',
-      filter: defaultBuildingFeatureFilter,
-      paint: {
-        'fill-color': [
-          'case',
-          ['boolean', ['feature-state', 'in_panel'], false],
-          selectedBuildingColor,
-          ['boolean', ['feature-state', 'hovered'], false],
-          '#132353',
-          '#1452e3',
-        ],
-        'fill-opacity': verifiedActive
-          ? [
-              'case',
-              ['boolean', ['feature-state', 'in_panel'], false],
-              0.15,
-              ['boolean', ['get', 'is_marked_as_correct'], false],
-              0.2,
-              0.08,
-            ]
-          : 0.08,
-      },
-    });
-
-    // Hatched overlay for buildings marked as correct
-    if (verifiedActive) {
-      map.addLayer({
-        id: LAYER_BDGS_SHAPE_MARKED_HATCH,
-        type: 'fill',
-        source: SRC_BDGS_SHAPES,
-        'source-layer': 'default',
-        filter: [
-          'all',
-          ...defaultBuildingFeatureFilter.slice(1),
-          ['==', 'is_marked_as_correct', true],
-        ],
-        paint: {
-          'fill-pattern': 'hatchGreen',
-          'fill-opacity': isSatellite ? 0.5 : 0.9,
-        },
-      });
-    }
-
-    // Polygon border
-    map.addLayer({
-      id: LAYER_BDGS_SHAPE_FILL,
-      type: 'line',
-      source: SRC_BDGS_SHAPES,
-      'source-layer': 'default',
-      filter: defaultBuildingFeatureFilter,
-      paint: {
-        'line-color': [
-          'case',
-          ['boolean', ['feature-state', 'in_panel'], false],
-          selectedBuildingColor,
-          ['boolean', ['feature-state', 'hovered'], false],
-          '#87d443',
-          '#1452e3',
-        ],
-        'line-width': 2.1,
-      },
-    });
-
-    // Points on the polygon source
-    map.addLayer({
-      id: LAYER_BDGS_SHAPE_POINT,
-      type: 'circle',
-      source: SRC_BDGS_SHAPES,
-      'source-layer': 'default',
-      filter: ['all', ['==', '$type', 'Point'], defaultBuildingFeatureFilter],
-      paint: {
-        'circle-radius': [
-          'case',
-          ['boolean', ['==', ['feature-state', 'hovered'], true]],
-          6,
-          5,
-        ],
-        'circle-stroke-color': verifiedActive
-          ? [
-              'case',
-              ['boolean', ['get', 'is_marked_as_correct'], false],
-              '#31e060',
-              '#ff0000',
-            ]
-          : '#ff0000',
-        'circle-stroke-width': 3,
-        'circle-color': [
-          'case',
-          ['boolean', ['feature-state', 'in_panel'], false],
-          '#31e060',
-          '#1452e3',
-        ],
-      },
-    });
-  };
-
-  const removeBuildings = (map: maplibregl.Map) => {
-    if (layers.buildings == 'point') {
-      removeBuildingsPoints(map);
-    }
-    if (layers.buildings == 'polygon') {
-      removeBuildingsShapes(map);
-    }
-  };
-
-  const removeBuildingsPoints = (map: maplibregl.Map) => {
-    LAYERS_BDGS_POINT_ALL.forEach((l) => {
-      if (map.getLayer(l)) {
-        map.removeLayer(l);
-      }
-    });
-    removeBuildingsSource(map);
-  };
-
-  const removeBuildingsShapes = (map: maplibregl.Map) => {
-    LAYERS_BDGS_SHAPE_ALL.forEach((l) => {
-      if (map.getLayer(l)) {
-        map.removeLayer(l);
-      }
-    });
-    removeBuildingsSource(map);
-  };
-
-  const removeBuildingsSource = (map: maplibregl.Map) => {
-    if (map.getSource(SRC_BDGS_POINTS)) {
-      map.removeSource(SRC_BDGS_POINTS);
-    }
-    if (map.getSource(SRC_BDGS_SHAPES)) {
-      map.removeSource(SRC_BDGS_SHAPES);
-    }
-  };
-
-  ///////////////////////////////////
-  ///////////////////////////////////
-  // Reports
-
-  const installReports = async (map: maplibregl.Map) => {
-    const darkColor = '#d64d00';
-    const lightColor = '#fcf5f4';
-
-    const zoomThreshold = 13;
-
-    if (map.getLayer(LAYER_REPORTS_CIRCLE))
-      map.removeLayer(LAYER_REPORTS_CIRCLE);
-    if (map.getLayer(LAYER_REPORTS_ICON)) map.removeLayer(LAYER_REPORTS_ICON);
-    if (map.getSource(SRC_REPORTS)) map.removeSource(SRC_REPORTS);
-
-    // add the icon if necessary
-    if (!map.hasImage('reportIcon')) {
-      const reportIconImg = await map.loadImage(reportIcon.src);
-      map.addImage('reportIcon', reportIconImg.data, { sdf: true });
-    }
-
-    map.addSource(SRC_REPORTS, {
-      type: 'vector',
-      tiles: [SRC_REPORTS_URL + '#' + Math.random()],
-      promoteId: 'id',
-    });
-
-    map.addLayer({
-      id: LAYER_REPORTS_SMALL_CIRCLES,
-      source: SRC_REPORTS,
-      'source-layer': 'default',
-      filter: getDefaultReportFilter(),
-      maxzoom: zoomThreshold,
-      type: 'circle',
-      paint: {
-        'circle-radius': 4,
-        'circle-color': darkColor,
-
-        'circle-stroke-color': lightColor,
-        'circle-stroke-width': 3,
-        'circle-stroke-opacity': 1,
-      },
-    });
-
-    map.addLayer({
-      id: LAYER_REPORTS_CIRCLE,
-      type: 'circle',
-      source: SRC_REPORTS,
-      'source-layer': 'default',
-      filter: getDefaultReportFilter(),
-      minzoom: zoomThreshold,
-      paint: {
-        'circle-radius': 15,
-        'circle-stroke-color': [
-          'case',
-          ['boolean', ['==', ['feature-state', 'hovered'], true]],
-          darkColor,
-          '#ffffff',
-        ],
-        'circle-stroke-width': 2,
-        'circle-color': [
-          'case',
-          ['boolean', ['feature-state', 'in_panel'], false],
-          darkColor,
-          lightColor,
-        ],
-      },
-    });
-
-    map.addLayer({
-      id: LAYER_REPORTS_ICON,
-      source: SRC_REPORTS,
-      'source-layer': 'default',
-      type: 'symbol',
-      filter: getDefaultReportFilter(),
-      minzoom: zoomThreshold,
-
-      layout: {
-        'icon-image': 'reportIcon',
-        'icon-size': 0.8,
-        'icon-allow-overlap': true,
-        'icon-ignore-placement': true,
-      },
-      paint: {
-        'icon-color': [
-          'case',
-          ['boolean', ['feature-state', 'in_panel'], false],
-          lightColor,
-          darkColor,
-        ],
-      },
-    });
-
-    setDisplayedReportFilters(map, displayedReportTags);
-  };
-
-  ///////////////////////////////////
-  ///////////////////////////////////
-  // Addresses
-
-  const installBAN = async (map: maplibregl.Map) => {
-    const certifiedColor = '#049c04';
-    const notCertifiedColor = '#777777';
-
-    if (map.getLayer(LAYER_BAN_POINT)) map.removeLayer(LAYER_BAN_POINT);
-    if (map.getLayer(LAYER_BAN_TXT)) map.removeLayer(LAYER_BAN_TXT);
-    if (map.getSource(SRC_BAN)) map.removeSource(SRC_BAN);
-
-    map.addSource(SRC_BAN, {
-      type: 'vector',
-      tiles: [SRC_BAN_URL],
-      minzoom: 10,
-      maxzoom: 14,
-
-      promoteId: 'id',
-    });
-
-    map.addLayer({
-      id: LAYER_BAN_POINT,
-      source: SRC_BAN,
-      'source-layer': 'adresses',
-      type: 'circle',
-      minzoom: 16,
-      paint: {
-        'circle-radius': 4,
-        'circle-stroke-color': '#ffffff',
-        'circle-stroke-width': 2,
-        'circle-color': [
-          'case',
-          ['==', ['get', 'certifie'], true],
-          certifiedColor,
-          notCertifiedColor,
-        ],
-      },
-    });
-
-    map.addLayer({
-      id: LAYER_BAN_TXT,
-      source: SRC_BAN,
-      'source-layer': 'adresses',
-      type: 'symbol',
-      minzoom: 16,
-      paint: {
-        // change text color if 'certifie' is true
-        'text-color': [
-          'case',
-          ['==', ['get', 'certifie'], true],
-          certifiedColor,
-          notCertifiedColor,
-        ],
-
-        'text-halo-color': '#ffffff',
-        'text-halo-width': 2,
-      },
-      layout: {
-        'text-font': ['Noto Sans Bold'],
-        'text-size': 14,
-
-        'text-field': [
-          'case',
-          ['has', 'suffixe'],
-          ['format', ['get', 'numero'], {}, ' ', {}, ['get', 'suffixe'], {}],
-          ['get', 'numero'],
-        ],
-        'text-ignore-placement': false,
-        'text-variable-anchor': ['bottom'],
-        'text-radial-offset': 1,
-      },
-    });
-  };
-
-  ///////////////////////////////////
-  ///////////////////////////////////
-  // Plots
-
-  const installPlots = useCallback((map: maplibregl.Map) => {
-    if (map.getLayer(LAYER_PLOTS_SHAPE)) map.removeLayer(LAYER_PLOTS_SHAPE);
-    if (map.getLayer(LAYER_PLOTS_TXT)) map.removeLayer(LAYER_PLOTS_TXT);
-    if (map.getSource(SRC_PLOTS)) map.removeSource(SRC_PLOTS);
-
-    map.addSource(SRC_PLOTS, {
-      type: 'vector',
-      tiles: [
-        process.env.NEXT_PUBLIC_API_BASE + '/plots/tiles/{x}/{y}/{z}.pbf',
-      ],
-      minzoom: 16,
-      maxzoom: 22,
-      promoteId: 'id',
-    });
-
-    map.addLayer({
-      id: LAYER_PLOTS_SHAPE,
-      source: SRC_PLOTS,
-      'source-layer': 'default',
-      type: 'line',
-      paint: {
-        'line-color': '#ea580c',
-        'line-opacity': 0.9,
-        'line-width': 2,
-      },
-    });
-
-    // Display the plot id in the middle of the plot
-    map.addLayer({
-      id: LAYER_PLOTS_TXT,
-      source: SRC_PLOTS,
-      'source-layer': 'default',
-      type: 'symbol',
-      layout: {
-        'text-field': ['get', 'plot_number'],
-        'text-size': 12,
-        'text-font': ['Open Sans Bold', 'Arial Unicode MS Bold'],
-        'text-offset': [0, 0],
-        'text-anchor': 'center',
-        'text-allow-overlap': false,
-      },
-      paint: {
-        'text-color': '#ea580c',
-        'text-halo-color': '#ffffff',
-        'text-halo-width': 1,
-      },
-    });
-  }, []);
 
   // When layers change, we rebuild the style and the layers
   useEffect(() => {
@@ -857,16 +147,4 @@ export const useMapLayers = ({
       installBuildings(map);
     }
   }, [reloadBuildings]);
-
-  const adsOperationsIcons = getADSOperationIcons();
-};
-
-export const getDefaultReportFilter = () => {
-  const defaultReportFilter: FilterSpecification = [
-    '==',
-    'pending',
-    ['get', 'status'],
-  ];
-
-  return defaultReportFilter;
 };
